+++ app/src/main/java/com/ch4vi/picsumgallery/App.kt
@@ -1,0 +1,7 @@
+package com.ch4vi.picsumgallery
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
